+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Add UnsafeEvaluateInvocationCheck to test for unsafe `unsafe_evaluate()` method calls.
 - Add SwChar16VectorEvaluateInvocationCheck to test for `sw:char16_vector.evaluate()` method calls.
 - Fix setting `magik.typing.index*Usages` internally in `magik-language-server`.
+- Add `do_not_translate` to SwModuleDefinitionGrammar and SwProductDefinitionGrammar.
 
 0.9.1 (2024-03-13)